moved getIframeContext function from src to /test/utils.ts
