+++ src/utils/common.ts
@@ -1,5 +1,4 @@
 import { KeyValueOption } from "../Common.types";
-import { Page } from "@playwright/test";
 
 // This function is used to check if an object is of type KeyValueOption
 export function isKeyValueOption(obj: unknown): obj is KeyValueOption {
@@ -12,16 +11,4 @@ export function uniqueSortedArray(array: string[], sortOrder?: "asc" | "desc") {
   return uniqueArray.sort((a, b) =>
     sortOrder === "asc" ? a.localeCompare(b) : b.localeCompare(a)
   );
-}
-
-// Helper function to wait for the iframe and switch to its context
-export async function getIframeContext(page: Page) {
-  const iframeElement = await page.waitForSelector(
-    'iframe[title="storybook-preview-iframe"]'
-  );
-  const iframe = await iframeElement.contentFrame();
-  if (!iframe) {
-    throw new Error("Unable to find or access the iframe content.");
-  }
-  return iframe;
 }
